+++ inc/deviceharddriveinjection.class.php
@@ -51,8 +51,10 @@ class PluginDatainjectionDeviceHardDriveInjection extends DeviceHardDrive implem
 
     public function connectedTo()
     {
+        /** @var array $CFG_GLPI */
+        global $CFG_GLPI;
 
-        return ["Computer"];
+        return $CFG_GLPI['itemdeviceharddrive_types'];
     }