+++ inc/deviceprocessorinjection.class.php
@@ -51,8 +51,10 @@ class PluginDatainjectionDeviceProcessorInjection extends DeviceProcessor implem
 
     public function connectedTo()
     {
+        /** @var array $CFG_GLPI */
+        global $CFG_GLPI;
 
-        return ["Computer"];
+        return $CFG_GLPI['itemdeviceprocessor_types'];
     }